search tests, part1
